Added RewardValue class and updated RewardsConverter
