+++ src/main/java/com/jpmorgan/RewardValue.java
@@ -1,0 +1,35 @@
+package com.jpmorgan;
+
+public class RewardValue {
+    private double cashValue;
+    private double milesValue;
+
+    private static final double MILES_TO_CASH_CONVERSION_RATE = 0.0035;
+
+    // Constructor that accepts a cash value
+    public RewardValue(double cashValue) {
+        this.cashValue = cashValue;
+        this.milesValue = cashValue / MILES_TO_CASH_CONVERSION_RATE;
+    }
+
+    // Constructor that accepts a miles value
+    public RewardValue(double milesValue, boolean isMiles) {
+        if (isMiles) {
+            this.milesValue = milesValue;
+            this.cashValue = milesValue * MILES_TO_CASH_CONVERSION_RATE;
+        } else {
+            this.cashValue = milesValue;
+            this.milesValue = milesValue / MILES_TO_CASH_CONVERSION_RATE;
+        }
+    }
+
+    // Method to get the cash value
+    public double getCashValue() {
+        return cashValue;
+    }
+
+    // Method to get the miles value
+    public double getMilesValue() {
+        return milesValue;
+    }
+}
